Fixed the issue with Mesos models not creating proper hashes.
Now, when a model has already run, the job id created will match that of a
previous run with the same function and same args. The issue was due to the
fact that the default hasher didn't take into account that Mesos model run
functions have additional args prepended to them.